Input system changes
