+++ Assets/Prefabs/Attacks.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b44e748bcd53b2a439db3033c887bcb3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
