+++ Assets/Prefabs/Attacks/BulletAttack.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 955e7b178406a4a41a7411fa8d80f4a4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
