+++ Assets/Prefabs/Attacks/PointLaserOrb.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 951f35511ed205f4ba6da23b58d57bd5, type: 3}
+  m_Name: PointLaserOrb
+  m_EditorClassIdentifier: 
+  projectilePrefab: {fileID: 712463402950107132, guid: 45118892dc3318447a90785c8b2ec39e, type: 3}
+  fireDelay: 3
+  target: {fileID: 0}
