+++ Assets/Prefabs/Attacks/PointLaserOrb.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0ef3ca281b2b8aa4289b8f24c83865e7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
