+++ Assets/Prefabs/Attacks/Shrapnel Bullet.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ce731468473b58b4988a683c822d85a6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
