+++ Assets/Prefabs/Attacks/SpawnFighters.asset
@@ -1,0 +1,16 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a40913829eeaa6d42a687028fb45439e, type: 3}
+  m_Name: SpawnFighters
+  m_EditorClassIdentifier: 
+  shipData: {fileID: 11400000, guid: d43fb925753bbbf46ba918e5b445efa0, type: 2}
+  fireDelay: 1
