+++ Assets/Prefabs/Attacks/SpawnFighters.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9ba436c7f169f704892d5151a8f12f31
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
