+++ Assets/Prefabs/EnemyScriptableObjects/Ships.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 969e665ff661d204f8fc1d980ffb81d5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
