+++ Assets/Prefabs/EnemyScriptableObjects/Ships/EnemyMothership.asset
@@ -1,0 +1,30 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6bb3a7adef543f4488c11cea31c320c0, type: 3}
+  m_Name: EnemyMothership
+  m_EditorClassIdentifier: 
+  ShipData:
+    ShipName: Mothership
+    Cost: 100
+    Health: 10
+    MaxHealth: 10
+    Speed: 0.1
+    StopDistance: 0.1
+    AggroRange: 100
+    BlocksMovement: 1
+    Weapons:
+    - {fileID: 11400000, guid: 0ef3ca281b2b8aa4289b8f24c83865e7, type: 2}
+    - {fileID: 11400000, guid: 9ba436c7f169f704892d5151a8f12f31, type: 2}
+    Modules: []
+    Buffs: []
+    StartingPos: {x: 0, y: 0}
+    ShipPrefab: {fileID: 4271728520197243408, guid: 33fbc513a6d90784cad0427918c1d263, type: 3}
