+++ Assets/Prefabs/EnemyScriptableObjects/Ships/EnemyMothership.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a06b30b53fd25ef43bad58b17671b938
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
