+++ Assets/Prefabs/PlayerScriptableObjects/Ships.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dd8680eab3863354fb5713a5b7636d78
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
