+++ Assets/Prefabs/PlayerScriptableObjects/Ships/FighterShip.asset
@@ -1,0 +1,29 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6bb3a7adef543f4488c11cea31c320c0, type: 3}
+  m_Name: FighterShip
+  m_EditorClassIdentifier: 
+  ShipData:
+    ShipName: Fighter
+    Cost: 50
+    Health: 10
+    MaxHealth: 10
+    Speed: 0.15
+    StopDistance: 0.1
+    AggroRange: 100
+    BlocksMovement: 0
+    Weapons:
+    - {fileID: 11400000, guid: 955e7b178406a4a41a7411fa8d80f4a4, type: 2}
+    Modules: []
+    Buffs: []
+    StartingPos: {x: 0, y: 0}
+    ShipPrefab: {fileID: 2473740277407861765, guid: 7b028babbc85d8649b7782681850be9c, type: 3}
