+++ Assets/Prefabs/PlayerScriptableObjects/Ships/FighterShip.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d43fb925753bbbf46ba918e5b445efa0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
