+++ Assets/Prefabs/PlayerScriptableObjects/Ships/Mothership.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b06386917a921804680d014d2b97a79a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
